Move args to utils
