+++ modules/genai_optimizations/benchmarks/milebench.py
@@ -20,6 +20,7 @@ from transformers import AutoProcessor
 
 from logging import getLogger
 from genai_opt import get_inputs_embeds
+from utils import add_visual_pruning_args
 
 
 logger = getLogger(__name__)
@@ -435,14 +436,6 @@ def get_model_class(model_name):
     else:
         error_msg = f"{model_name} is not supported."
         raise NotImplementedError(error_msg)
-
-
-def add_visual_pruning_args(parser):
-    group = parser.add_argument_group("Visual Token Pruning Arguments")
-    group.add_argument("--enable_visual_pruning", action="store_true", help="Enable visual token pruning")
-    group.add_argument("--num_keep_tokens", type=int, default=128, help="Number of visual tokens to keep")
-    group.add_argument("--theta", type=float, default=0.5, help="Balance factor for diversity vs relevance")
-    return parser
 
 
 if __name__ == "__main__":
